requirements.txt
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.4.2
GitPython>=3.1.40